fix(tests): keep the shared test worker loadable outside the vitest pool
The capability fixture files mixed harness Durable Object classes with
their cloudflare:test-importing drivers, so worker.ts transitively pulled
cloudflare:test — a module that only exists inside the vitest pool. The
React project boots that worker under wrangler unstable_dev, so its global
setup failed and every browser test burned its full retry budget, which is
what pushed CI's agents:test past the job timeout. Harness classes stay in
tests/capabilities/ (worker-safe, documented rule); withCapabilityHarness
and withMcpHarness join the other pool-only drivers in tests/shared/.
